+++ README.md
@@ -3,6 +3,18 @@
 
 ## 链接🔗
 
-https://mimotion.vercel.app/
+https://mimotion.vercel.app/mimotion
+
+参数：
+
+u : username
+
+p : password
+
+s : step
+
+例如（不能直接点击，需要修改参数）：
+
+https://mimotion.vercel.app/mimotion?u=18888888888&p=123456789&s=20090
 
 ## 预览
